arm: bsp: refactor NXP MRU driver for reuse across SoCs
Move the MRU driver from bsp/s32z to drivers/arm to make it available
for other platforms that integrate the MRU peripheral. This refactoring
separates generic driver logic from SoC-specific details.
Platform-specific parts, such as interrupt count and the factory method
for driver instantiation, remain under BSP specific directory.

No functional change intended.
